Re-implemented eagerize/eagerize using protocols.
This is 3x faster then
cond, and is now open for extenssion, so if users have custom container
types, they can simply extend them to support eagerize and it will work
fine with special.

For some reason, it doesn't work to extend object arrays for subtypes,
so I had to implement it as part of the java.lang.Object case.